Isolate the legacy input_required shim into its own module
Server no longer carries the fulfilment loop inline: it lazily holds
one LegacyInputRequiredShim (legacyInputRequiredShim.ts) constructed
against a narrow host contract — knobs, the resolved capability view,
the requestState verify runner, and the three capability-check-free
sender cores — and the seam delegates to it in one line. The shared
embedded-request validation moves with it (the modern capability check
imports it back), withRequestStateValue moves to core-internal next to
requestStateAccessor, and server.ts shrinks by ~300 lines. Behavior
unchanged; package-internal, not exported from the index.
